DS 2710 (#14)
* Update widget formatting to be consistent [revdep skip]

Currently the flipTextAnalysis outputs are not consistent with
other table outputs such as Regression outputs. This commit
updates the output tables to be in the same format as the
flipRegression outputs

* Update documentation to roxygen2 v7.0.2 [revdep skip]

* Update unit tests for new footer format [revdep skip]

* Remove old function and tidy up textanalysis.css

The function for printing the CV metrics in TextClassifier has been
removed since it is computed instead in the print method in
flipTextAnalysis. Also cleaned up the css file for textanalysis.css,
footer settings were defined twice

* Modify css for Displayr output [revdep skip]

Due to slight differences between css implementation locally
vs Displayr, the css files needed to be updated to account for
settings being overrided for the TextClassifier CV table. Also
changed the padding settings for wordbag output

* Reduce footprint TextClassifierWidget test [revdep skip]

Previously, TextClassifierWidget was tested by fully specifying the R output
variables for the Tom Cruise Data with an existing categorization. The full
output has been trimmed for the examples and an inline comment added in the
test source file

* Modify TextAnalysis Widget to allow dynamic footer size