DS-2703 (#18)
* DS-2703 Allow Labels to handle edge case [revdep skip]

In the new Stack data option in flipRegression, a question with a multi
variable set structure is reduced to a single vector. However, the
original Question does not have a name attribute but it does have a
label. This can cause unhelpful output with the Labels function that
expects a name to exist if the label exists.

* DS-2703 Add nloptr library to travis

Add libnlopt-dev to travis.yml to ensure nloptr can build as its a
dependency for flipChoice that is required for unit tests.

* DS-2703 Add unit test for desired behaviour [revdep skip]

Co-authored-by: flipDevTools <opensource@displayr.com>